+++ src/lib/config/SliderConfig.svelte
@@ -27,6 +27,7 @@
     onValueChange={values => onValueChange(values[0])}
     min={range.start}
     max={range.end}
+    step={type === 'int32' ? 1 : 0.01}
     class="flex-grow relative flex items-center group"
 >
     <div class="flex-grow bg-gray-900 h-2 rounded-full">